// src/types/clip-tag-enum.ts
export enum ClipTag {
	LAYUP = "#layup",
	FLOATER = "#floater",
	HOOK = "#hook",
	PUT_BACK = "#putback",
	ALLEY_OOP = "#alleyoop",
	CIRCUS = "#circus",
	MID_RANGE = "#midrange",
	LONG_RANGE = "#longrange",
	DEEP = "#deep",
	BANK_SHOT = "#bankshot",
	FADE_AWAY = "#fadeaway",
	CATCH_AND_SHOOT = "#catchandshoot",
	OFF_THE_DRIBBLE = "#offthedribble",
	AND1 = "#and1",
	HIGHLIGHT = "#highlight",
	GAME_WINNER = "#gamewinner",
	PICK_AND_ROLL = "#pickandroll",
	PICK_AND_POP = "#pickandpop",
	NO_LOOK = "#nolook",
	CONTESTED = "#contested",
	OPEN = "#open",
}